Add files via upload
*** Interface of final app is in Romanian language, code isn't (perphaps some variable values in the source code were written in RO also).

Compiled release of this application. This app is only an approach of solving an University project and I am not planning to continue it as it already served it's purpose. It's a calculation interpretation for every student having certain individual values to calculate in a simulation of Energy Management degree materia. What exactly it does?

There was a project that needed to be solved with certain terms :
- Manage/simulate a power plant with a certain number of consumers, rural and urban. Every student has it's own case of calculations (by numbers). 
- Manage an industrial warehouse with a certain number of engines of different power, size and weight, maintain a drawn scheme for specified warehouse, with all of electric and non electric dimensions and values, including length and width in metric system, every engine's nominal power.
- Select wires accordingly to power needs and also specify it length to calculate current intensity and impedance per every metre. 
- Adjust power plant electrical and non-electrical properties to correspond with rural, urban consumers and with industrial warehouse power needs.

All of these parameters required a mass of calculations and improvised drawing. My application was an attempt to make these calculations by limited clicks and make project and make the project faster.

What was used for this application:
- Undo-redo command pattern (it's like rudimentary implementation of default Windows Paint application, but single purposed only) for adding and removing engines in industrial warehouse image simulation (see info below).
- BMP image library asset to create and save image of the power plant block scheme with rural and urban consumers per every student case (a certain variable corresponding for every student).
- Mahapps Metro library for making user interface look friendly.

*** This app doesn't have any other useful purpose for third faces, except how "Undo-redo command pattern" works in this case but dissecting code would be hard for those who never coded in WPF C# and there aren't much of comments on every code line. 
It's my first project of my self-learning programming course, you can download the archive and launch it to see what it looks like.

*** It's a work of a novice coder and I am just learning to get better.